Cart service completed

+++ src/main/java/com/example/demo/model/Product.java
@@ -7,6 +7,7 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 @Getter
@@ -18,7 +19,7 @@ import java.util.List;
 @Table(name="products")
 public class Product {
 
-    public Product(String name, String description, double price, int quantity, String brand, Category category) {
+    public Product(String name, String description, BigDecimal price, int quantity, String brand, Category category) {
         this.name = name;
         this.description = description;
         this.price = price;
@@ -32,7 +33,7 @@ public class Product {
     private Long id;
     private String name;
     private String description;
-    private double price;
+    private BigDecimal price;
     private int quantity;
     private String brand;
